requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pyscopg2==66.0.2
pytz==2023.3
sqlparse==0.4.4
